Refactor room pages
