+++ Web/src/Routes.tsx
@@ -10,8 +10,6 @@ const JoinRoomPage = lazy(() => import('components/join-room'));
 const BalloonShake = lazy(() => import('components/games/BalloonShake'));
 const PandaSequence = lazy(() => import('components/games/PandaSequence'));
 const MentalSums = lazy(() => import('components/games/MentalSums'));
-const HostRoomPage = lazy(() => import('components/create-room/HostRoom'));
-const WaitingRoom = lazy(() => import('components/waiting'));
 
 const comm = new PhoenixComm();
 
@@ -21,11 +19,9 @@ const RoutesWithCommContext: React.FC = () => (
       <Route path="/new" component={CreateRoomPage} />
       <Route exact path="/join" component={JoinRoomPage} />
       <Route path="/join/:id" component={JoinRoomPage} />
-      <Route path="/room" component={HostRoomPage} />
       <Route exact path="/balloon-game" component={BalloonShake} />
       <Route exact path="/panda-sequence" component={PandaSequence} />
       <Route exact path="/sums-game" component={MentalSums} />
-      <Route exact path="/participant-room" component={WaitingRoom} />
     </Switch>
   </CommContext.Provider>
 );
