+++ Web/src/components/create-room/HostRoom.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import BigButton from 'components/common/BigButton';
 import CommContext from 'components/room/comm/CommContext';
 import CommonRoom, { FinishedComponentProps } from 'components/room/CommonRoom';
-import { resultsExist } from 'components/room/comm/Comm';
+import { resultsExist, CommAttributes } from 'components/room/comm/Comm';
 import NumPlayers from './NumPlayers';
 import SocialShare from './SocialShare';
 import QrCode from './QrCode';
@@ -171,8 +171,9 @@ const HostRoomLobby: React.FC<FinishedComponentProps> = ({
   );
 };
 
-const HostRoomPage: React.FC = () => (
+const HostRoomPage: React.FC<{ commHooks: CommAttributes }> = ({ commHooks }) => (
   <CommonRoom
+    commHooks={commHooks}
     FinishedComponent={HostRoomLobby}
   />
 );
