+++ Web/src/components/room/CommonRoom.tsx
@@ -2,26 +2,21 @@ import React, {
   useState, useContext, useEffect, lazy,
 } from 'react';
 import CommContext from 'components/room/comm/CommContext';
-import useCommHooks from 'components/room/comm/useCommHooks';
 import Loading from 'components/common/Loading';
 import GameState from './comm/GameState';
 import Game from './Games';
-import { CommError } from './comm/Errors';
-import { ResultMap } from './comm/Comm';
+import { CommAttributes } from './comm/Comm';
 
 const BalloonShake = lazy(() => import('components/games/BalloonShake'));
 const MentalSums = lazy(() => import('components/games/MentalSums'));
 const PandaSequence = lazy(() => import('components/games/PandaSequence'));
 
-export interface FinishedComponentProps {
-  results: ResultMap | null;
-  room: string | null;
-  error: CommError | null;
-  users: string[];
+export interface FinishedComponentProps extends CommAttributes {
   game: Game;
 }
 
 interface CommonRoomProps {
+  commHooks: CommAttributes;
   FinishedComponent: React.FC<FinishedComponentProps>;
   NoHostComponent?: React.FC;
 }
@@ -38,6 +33,7 @@ const GameComponent: React.FC<{ game: Game }> = ({ game }) => (
 );
 
 const CommonRoom: React.FC<CommonRoomProps> = ({
+  commHooks,
   FinishedComponent,
   NoHostComponent = () => <p>No Host :(</p>,
 }) => {
@@ -47,9 +43,7 @@ const CommonRoom: React.FC<CommonRoomProps> = ({
   // general hook to disconnect host from room when he leaves.
   useEffect(() => () => comm.leaveRoom(), [comm]);
 
-  const {
-    hostMeta, room, error, users, results,
-  } = useCommHooks(comm);
+  const { hostMeta } = commHooks;
 
   useEffect(() => {
     if (hostMeta === null) return;
@@ -65,9 +59,8 @@ const CommonRoom: React.FC<CommonRoomProps> = ({
       {hostMeta.state === GameState.FINISHED
         && (
           <FinishedComponent
-            {... {
-              results, room, users, error, game,
-            }}
+            {...{ ...commHooks }}
+            game={game}
           />
         )}
       {hostMeta.state === GameState.ONGOING
